requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
idna==3.4
protobuf==3.20.3
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4.1
urllib3==1.26.15
